+++ src/components/info-window.tsx
@@ -1,6 +1,7 @@
 /* eslint-disable complexity */
 import React, {
   CSSProperties,
+  FunctionComponent,
   PropsWithChildren,
   ReactNode,
   useEffect,
@@ -34,7 +35,9 @@ export type InfoWindowProps = Omit<
 /**
  * Component to render an Info Window with the Maps JavaScript API
  */
-export const InfoWindow = (props: PropsWithChildren<InfoWindowProps>) => {
+export const InfoWindow: FunctionComponent<
+  PropsWithChildren<InfoWindowProps>
+> = props => {
   const {
     // content options
     children,
@@ -169,7 +172,7 @@ export const InfoWindow = (props: PropsWithChildren<InfoWindowProps>) => {
 
   // ## open info window when content and map are available
   const map = useMap();
-  useEffect(() => {
+  useDeepCompareEffect(() => {
     // `anchor === null` means an anchor is defined but not ready yet.
     if (!map || !infoWindow || anchor === null) return;